+++ src/main/java/de/omnikryptec/libapi/opengl/texture/GLTexture2D.java
@@ -23,9 +23,9 @@ import de.omnikryptec.resource.TextureConfig;
 import de.omnikryptec.resource.TextureData;
 
 public class GLTexture2D extends GLTexture {
-
+    
     private final TextureData data;
-
+    
     public GLTexture2D(final TextureData texture, final TextureConfig config) {
         super(GL11.GL_TEXTURE_2D);
         this.data = texture;
@@ -33,20 +33,20 @@ public class GLTexture2D extends GLTexture {
         OpenGLUtil.loadTexture(texture);
         OpenGLUtil.configureTexture(config);
     }
-
+    
     @Override
-    public float getWidth() {
+    public int getWidth() {
         return this.data.getWidth();
     }
-
+    
     @Override
-    public float getHeight() {
+    public int getHeight() {
         return this.data.getHeight();
     }
-
+    
     @Override
     public boolean requiresInvertedVifDrawn2D() {
         return true;
     }
-
+    
 }
